Mentorship Dropdown menu corrections
