+++ playwright-tests/utils/fixtures.ts
@@ -3,7 +3,7 @@ import { test as base } from '@playwright/test';
 import { BasePage } from '@pages/base.page';
 import { HomePage } from '@pages/home.page';
 import { LandingPage } from '@pages/landing.page';
-import { MentorshipDropdownPage } from '@pages/mentorship-dropdown.page';
+import { MentorshipDropdownPage } from '@pages/mentorshipdropdown.page';
 
 // Declare the types of your fixtures.
 type PageObjectFixtures = {
@@ -11,7 +11,6 @@ type PageObjectFixtures = {
   basePage: BasePage;
   homePage: HomePage;
   mentorshipDropdownPage: MentorshipDropdownPage;
-
 };
 
 export const test = base.extend<PageObjectFixtures>({
@@ -32,5 +31,4 @@ export const test = base.extend<PageObjectFixtures>({
     const mentorshipDropdownPage = new MentorshipDropdownPage(page);
     await use(mentorshipDropdownPage);
   },
-
 });
